--- packages/apollo-shared/src/ChangeManager/changes/index.ts
@@ -1,13 +1,0 @@
-export * from './AddAssemblyAndFeaturesFromFileChange'
-export * from './AddAssemblyFromFileChange'
-export * from './AddFeatureChange'
-export * from './AddFeaturesFromFileChange'
-export * from './CopyFeatureChange'
-export * from './DeleteAssemblyChange'
-export * from './DeleteFeatureChange'
-export * from './DeleteUserChange'
-export * from './FeatureAttributeChange'
-export * from './LocationEndChange'
-export * from './LocationStartChange'
-export * from './TypeChange'
-export * from './UserChange'
