--- packages/apollo-shared/src/ChangeManager/index.ts
@@ -1,36 +1,0 @@
-import {
-  AddAssemblyAndFeaturesFromFileChange,
-  AddAssemblyFromFileChange,
-  AddFeatureChange,
-  AddFeaturesFromFileChange,
-  CopyFeatureChange,
-  DeleteAssemblyChange,
-  DeleteFeatureChange,
-  DeleteUserChange,
-  FeatureAttributeChange,
-  LocationEndChange,
-  LocationStartChange,
-  TypeChange,
-  UserChange,
-} from './changes'
-
-export const changes = {
-  AddAssemblyAndFeaturesFromFileChange,
-  AddAssemblyFromFileChange,
-  AddFeatureChange,
-  AddFeaturesFromFileChange,
-  CopyFeatureChange,
-  DeleteAssemblyChange,
-  DeleteFeatureChange,
-  DeleteUserChange,
-  FeatureAttributeChange,
-  LocationEndChange,
-  LocationStartChange,
-  TypeChange,
-  UserChange,
-}
-
-export * from './changes'
-export * from './ChangeTypes'
-
-export * from './changes/abstract'
