+++ packages/apollo-shared/src/Changes/AddAssemblyAndFeaturesFromFileChange.ts
