+++ packages/apollo-shared/src/Changes/AddAssemblyFromFileChange.ts
