+++ packages/apollo-shared/src/Changes/CopyFeatureChange.ts
