+++ packages/apollo-shared/src/Changes/DeleteFeatureChange.ts
