+++ packages/apollo-shared/src/Changes/LocationEndChange.ts
