+++ packages/apollo-shared/src/Changes/abstract/AssemblySpecificChange.ts
