+++ packages/apollo-shared/src/Changes/abstract/index.ts
