+++ packages/apollo-shared/src/Validations/ValidationSet.ts
@@ -1,7 +1,7 @@
 import { FeatureDocument, UserDocument } from 'apollo-schemas'
 import { ClientSession, Model } from 'mongoose'
 
-import { Change, ClientDataStore } from '../ChangeManager/changes/abstract'
+import { Change, ClientDataStore } from '../Changes/abstract'
 import { Context, Validation, ValidationResult } from './Validation'
 
 export class ValidationResultSet {
