+++ packages/apollo-shared/src/index.ts
@@ -1,3 +1,4 @@
-export * from './ChangeManager'
+export * from './Changes'
+export * from './Changes/abstract'
 export * from './Validations'
 export * from './Common'
